src/service-setting-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
azure-cosmos==4.5.1
redis==5.0.1
python-multipart==0.0.6
